+++ changelog.md
@@ -1,2 +1,11 @@
+# 1.1.0
+- Layer names in 'Edit Group' menu
+- Preserve layer names after copying the level
+- More <cy>keybinds</c>
+- Save object
+- New mod logo (thanks, Rue)
+
+
 # 1.0.0
-- Initial release
+- Named layers in editor
+- List of named layers with additional info